Create ExplorerPatcher_WARNING_20230224
I don't think this program is malicious, but it caused a LOT of grief for me! I wanted to get two rows on my Taskbar, and found a tech site that recommended this app.
I ran it and it appears to have killed all processes, and then got stuck in an unstopable loop. I was able to start Task Manager and get a command prompt window open in Adminstrator mode, but couldn't find the process to kill it.
I am running Windows 11 Pro recently updated Version 22H2 OS Build 22621.1343, and a Start Menu replacement called Start 11. I have been using their Start Menu replacement since Windows 8, and this is the first time it seems to have a major problem with a Windows update.
Luckily I had made a Restore Point before I ran this utility, and was able to get to System Restore by issuing the command "start sysdm.cpl" from the command prompt.
It took over a half hour for it to complete, and on the first reboot there were all kinds of problems with startup apps, and even Notepad wouldn't start. Luckily a reboot cleared that up, but I still don't have a 2 row Taskbar, the icons are large and cut off (I know how to reduce their size, but then the System Tray disappears), and Maximized windows have a noticeable gap between the bottom of the window and the Taskbar. It's ugly, but at least it's working decently!
Hope this can be fixed, and someone can let me know when it is.